src: removes more code from cli.js Tests added, after refactor. Connects to #211
