+++ bin/cli.js
@@ -1,8 +1,6 @@
 'use strict';
 
-const reader = require('../lib/file-reader.js');
 const writer = require('../lib/file-writer.js');
-const remoteLicense = require('../lib/remote-license');
 const project = require('../lib/project');
 const licenseDependency = require('../lib/license-dependency');
 
@@ -11,19 +9,7 @@ const licenseDependency = require('../lib/license-dependency');
 // print warnings and create html in case needed.
 function run (options) {
   if (project.hasNodeModules(options.directory)) {
-    let mappings = [];
-    if (options.nameMap) {
-      if (options.nameMap.startsWith('http')) {
-        mappings = remoteLicense.fetch(options.nameMap);
-      } else {
-        mappings = reader.readAsJson(options.nameMap);
-      }
-      if (mappings === null) {
-        console.error('Could not find name map file: ', options.nameMap);
-        process.exit(3);
-      }
-    }
-    licenseDependency.initNameMapper(mappings);
+    licenseDependency.initNameMapper(options);
     if (options.merge) {
       writer.mergeXmls(options);
       return;
