+++ lib/license-dependency.js
@@ -8,12 +8,26 @@ const writer = require('./file-writer');
 const messages = require('./messages');
 const versionHandler = require('./version-handler');
 const dual = require('./dual');
+const remoteLicense = require('./remote-license');
 
 let canonicalNameMapper;
 let dependencyLicenseFiles = [];
 
-const initNameMapper = (mappings) => {
+const initNameMapper = (options) => {
+  let mappings = [];
+  if (options.nameMap) {
+    if (options.nameMap.startsWith('http')) {
+      mappings = remoteLicense.fetch(options.nameMap);
+    } else {
+      mappings = reader.readAsJson(options.nameMap);
+    }
+    if (mappings === null) {
+      console.error('Could not find name map file: ', options.nameMap);
+      process.exit(3);
+    }
+  }
   canonicalNameMapper = require('./canonical-name.js')(mappings);
+  return mappings;
 };
 
 // Shows warning messages if needed.
